Fix Microsoft Teams notification in i3
Microsoft Teams client on linux does not use native notifications,
instead it opens each notification as a seperate window. In tiling
window managers, such as i3, this causes half of the screen to be taken
up by the notification popup. To fix this we set the notification
windows to floating.