Add gitignore to fix yadm auto-complete
`yadm add` would hang while tab completing since it scanned the entire
home catalog for files. To avoid this we need to ignore some paths. I
tried doing a reverse gitignore setup but that was not as easy as I
thought. Instead we simply ignore paths that are problematic.

Fix .gitignore